Legacy theme and MUI theme separated into their own files
- Updated index.ts accordingly so theme consumers don't need changes
