+++ tdesign/src/themes/legacyTheme.ts
@@ -1,6 +1,6 @@
 import { SxProps } from "@material-ui/system";
 import { Theme } from "@material-ui/core/styles/createMuiTheme";
-import { defaultTheme } from "./defaultTheme";
+import { defaultTheme } from "./muiTheme";
 
 /* 
 Pre-April 2021, different areas used different styles provided by different files 
@@ -14,6 +14,36 @@ _app: import "@csstools/normalize.css";
       import "../styles/base.css";
       import "@splitgraph/design/css/base.css";
 */
+
+/*
+  Use the @rebass/preset theme as a starting point.
+
+  Taken from:
+    https://github.com/rebassjs/rebass/blob/master/packages/preset/src/index.js
+*/
+
+// import prismTheme from "@theme-ui/prism/presets/shades-of-purple.json";
+
+// const prismTheme = {
+//   color: "#9EFEFF",
+//   backgroundColor: "#2D2A55",
+//   ".changed": { color: "rgb(255, 238, 128)" },
+//   ".deleted": { color: "rgba(239, 83, 80, 0.56)" },
+//   ".inserted": { color: "rgb(173, 219, 103)" },
+//   ".comment": { color: "rgb(179, 98, 255)", fontStyle: "italic" },
+//   ".punctuation": { color: "rgb(255, 255, 255)" },
+//   ".constant": { color: "rgb(255, 98, 140)" },
+//   ".string,.url": { color: "rgb(165, 255, 144)" },
+//   ".variable": { color: "rgb(255, 238, 128)" },
+//   ".number,.boolean": { color: "rgb(255, 98, 140)" },
+//   ".attr-name": { color: "rgb(255, 180, 84)" },
+//   ".keyword,.operator,.property,.namespace,.tag,.selector,.doctype": {
+//     color: "rgb(255, 157, 0)",
+//   },
+//   ".builtin,.char,.constant,.function,.class-name": {
+//     color: "rgb(250, 208, 0)",
+//   },
+// };
 
 export const prismTheme = {
   color: "#e0ffff",
@@ -86,6 +116,49 @@ export const tocStyles = {
   ".toc-level": {},
 
   ".toc-link": {},
+};
+
+export const marketingTheme = {
+  styles: {
+    pre: {
+      ...prismTheme,
+      padding: "1ch",
+      [defaultTheme.breakpoints.down("sm")]: {},
+      fontSize: "0.8rem",
+      overflowX: "auto",
+      MsOverflowStyle: "none",
+      backgroundColor: "primary",
+      ".mdx-marker": {
+        display: "block",
+        borderLeft: `.25em solid ${prismTheme[".punctuation"].color}`,
+      },
+    },
+    inlineCode: {
+      ...prismTheme,
+      [defaultTheme.breakpoints.up("sm")]: {
+        minWidth: "initial",
+      },
+      paddingTop: 0,
+      paddingBottom: 0,
+      minHeight: "1rem",
+      display: "inline-flex",
+      alignContent: "center",
+      overflowX: "auto",
+      backgroundColor: prismTheme.color,
+      // color: "red",
+    },
+    code: {
+      backgroundColor: "primary.main",
+      fontFamily: "monospace",
+      fontSize: "inherit",
+      span: {
+        // color: "red",
+      },
+      ".comment": {
+        color: "#f4c1c0",
+      },
+    },
+  },
 };
 
 export const makeDefaultTheme = () => ({
